Fix unit tests for pytest_ansible compatibility (#2858)
SUMMARY
Fix unit tests to work correctly when run using the pytest_ansible plugin. The pytest_ansible plugin isn't as aggressive as the ansible_test plugin when it comes to rewriting import paths.
This is a follow-up to #2816.
Changes include:

Move shared test code to tests/unit/utils to avoid namespace pollution
Use fully qualified imports throughout test files
Update unit test requirements
Clean up deprecated imports

ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
tests
ADDITIONAL INFORMATION
This change ensures that unit tests can be run using pytest_ansible without import errors. The tests were previously relying on import path rewriting that was specific to the ansible_test plugin.
