+++ tests/unit/plugins/module_utils/conftest.py
@@ -6,14 +6,14 @@
 import json
 import sys
 import warnings
+from collections.abc import MutableMapping
 from io import BytesIO
 
 import pytest
 
 import ansible.module_utils.basic
 import ansible.module_utils.common
-from ansible.module_utils._text import to_bytes
-from ansible.module_utils.common._collections_compat import MutableMapping
+from ansible.module_utils.common.text.converters import to_bytes
 
 
 @pytest.fixture(name="stdin")
