+++ tests/unit/plugins/modules/utils.py
@@ -3,7 +3,7 @@ import unittest
 from unittest.mock import patch
 
 from ansible.module_utils import basic
-from ansible.module_utils._text import to_bytes
+from ansible.module_utils.common.text.converters import to_bytes
 
 
 def set_module_args(args):
