+++ tests/unit/requirements.txt
@@ -4,5 +4,6 @@ boto3
 
 placebo
 # For testing event source plugins
+aiobotocore
 asyncmock
 pytest-asyncio
