+++ tests/unit/utils/event.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+# This file is part of Ansible
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+import asyncio
+from typing import Any
+
+
+class ListQueue(asyncio.Queue[Any]):
+    def __init__(self) -> None:
+        self.queue: list[Any] = []
+
+    async def put(self, item: Any) -> None:
+        self.queue.append(item)
+
+    def put_nowait(self, item: Any) -> None:
+        self.queue.append(item)
